dont vote on the code review label
for now we are still developign the review job
so we want to vote 0 for all review comments until
we are sure the infra is stable.

in the future we will start by voting 1 when there are no
critical issues found and reserve -1 for critical detections
only. later we may expand that but its imporatnt that
the signal to noise ratio be high.

Signed-off-by: Sean Mooney <sean@seanmooney.info>